Refactored event-service
